+++ libs/core/multi-combobox/multi-combobox.component.ts
@@ -596,13 +596,11 @@ export class MultiComboboxComponent<T = any> extends BaseMultiCombobox<T> implem
     _toggleSelection(item: SelectableOptionItem, fromTokenCloseClick = false): void {
         const selectedSuggestions = this._selectedSuggestions();
         const idx = getTokenIndexByIdlOrValue(item, selectedSuggestions);
-        if (idx === -1) {
-            this._selectedSuggestions.set([...selectedSuggestions, item]);
-        } else {
-            this._selectedSuggestions.set(selectedSuggestions.filter((_, i) => i !== idx));
-        }
+        const updatedSelectedSuggestions =
+            idx === -1 ? [...selectedSuggestions, item] : selectedSuggestions.filter((_, index) => index !== idx);
 
         item.selected = !item.selected;
+        this._setSelectedSuggestions(updatedSelectedSuggestions.map(({ value }) => value));
 
         this._propagateChange(fromTokenCloseClick);
 
@@ -652,11 +650,15 @@ export class MultiComboboxComponent<T = any> extends BaseMultiCombobox<T> implem
         if (event) {
             event.preventDefault();
         }
-        const optionItem = this._flatSuggestions().find((s) => s.value === token.value);
-        if (optionItem) {
-            this._toggleSelection(optionItem, true);
-            this._rangeSelector.reset();
+        const flatSuggestions = this._flatSuggestions();
+        const optionItemIndex = getTokenIndexByIdlOrValue(token, flatSuggestions);
+
+        if (optionItemIndex === -1) {
+            return;
         }
+
+        this._toggleSelection(flatSuggestions[optionItemIndex], true);
+        this._rangeSelector.reset();
     }
 
     /** @hidden */
@@ -724,7 +726,7 @@ export class MultiComboboxComponent<T = any> extends BaseMultiCombobox<T> implem
      * Handle dialog dismissing, closes popover and sets backup data.
      */
     _dialogDismiss(backup: SelectableOptionItem[]): void {
-        this._selectedSuggestions.set([...backup]);
+        this._setSelectedSuggestions(backup.map(({ value }) => value));
         this._setInputText('');
         this._showList(false);
         this.selectedShown.set(false);